# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CustomerDetail": {
        "properties": {
          "band": {
            "anyOf": [
              {
                "enum": [
                  "hot",
                  "warm",
                  "cold"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Band"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identities": {
            "items": {
              "$ref": "#/components/schemas/Identity"
            },
            "title": "Identities",
            "type": "array"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "leads": {
            "items": {
              "$ref": "#/components/schemas/LeadOut"
            },
            "title": "Leads",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "open_tasks": {
            "title": "Open Tasks",
            "type": "integer"
          },
          "opted_out": {
            "title": "Opted Out",
            "type": "boolean"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "profile": {
            "additionalProperties": true,
            "title": "Profile",
            "type": "object"
          },
          "profile_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Updated At"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "country",
          "language",
          "tags",
          "owner",
          "band",
          "opted_out",
          "last_seen_at",
          "identities",
          "profile",
          "profile_updated_at",
          "leads",
          "open_tasks"
        ],
        "title": "CustomerDetail",
        "type": "object"
      },
      "CustomerPage": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CustomerSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data",
          "next_cursor"
        ],
        "title": "CustomerPage",
        "type": "object"
      },
      "CustomerPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "CustomerPatch",
        "type": "object"
      },
      "CustomerSummary": {
        "properties": {
          "band": {
            "anyOf": [
              {
                "enum": [
                  "hot",
                  "warm",
                  "cold"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Band"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "opted_out": {
            "title": "Opted Out",
            "type": "boolean"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "country",
          "language",
          "tags",
          "owner",
          "band",
          "opted_out",
          "last_seen_at"
        ],
        "title": "CustomerSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Identity": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_primary": {
            "title": "Is Primary",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "value",
          "is_primary"
        ],
        "title": "Identity",
        "type": "object"
      },
...
      "MergeIn": {
        "properties": {
          "keep_id": {
            "format": "uuid",
            "title": "Keep Id",
            "type": "string"
          },
          "merge_id": {
            "format": "uuid",
            "title": "Merge Id",
            "type": "string"
          }
        },
        "required": [
          "keep_id",
          "merge_id"
        ],
        "title": "MergeIn",
        "type": "object"
      },
...
      "ReassignIn": {
        "properties": {
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          }
        },
        "required": [
          "owner_id"
        ],
        "title": "ReassignIn",
        "type": "object"
      },
...
      "TimelineEntry": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "message",
              "activity"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id",
          "at",
          "data"
        ],
        "title": "TimelineEntry",
        "type": "object"
      },
      "TimelinePage": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntry"
            },
            "title": "Data",
            "type": "array"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "required": [
          "data",
          "next_offset"
        ],
        "title": "TimelinePage",
        "type": "object"
      },
...
      "UserRef": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "UserRef",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DealerAI OS",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/customers": {
      "get": {
        "operationId": "list_customers_v1_customers_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "hot",
                    "warm",
                    "cold"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Band"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "default": "",
              "title": "Country",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "default": "",
              "title": "Tag",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Customers",
        "tags": [
          "customers"
        ]
      }
    },
    "/v1/customers/merge": {
      "post": {
        "description": "Two records, one customer. Not undoable — the audit row is what a repair\nwould start from, and the old id keeps answering with where it went.",
        "operationId": "merge_customers_v1_customers_merge_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Customers",
        "tags": [
          "customers"
        ]
      }
    },
    "/v1/customers/{customer_id}": {
      "get": {
        "operationId": "get_customer_v1_customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Customer",
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "description": "Anything typed here is a person's answer, stamped `human` — which is what\nstops the next AI run overwriting it.",
        "operationId": "edit_customer_v1_customers__customer_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Customer",
        "tags": [
          "customers"
        ]
      }
    },
    "/v1/customers/{customer_id}/reassign": {
      "post": {
        "description": "The customer, their open conversations, their open leads and their open\ntasks, in one transaction — and a line in every thread that moved.",
        "operationId": "reassign_customer_v1_customers__customer_id__reassign_post",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reassign Customer",
        "tags": [
          "customers"
        ]
      }
    },
    "/v1/customers/{customer_id}/timeline": {
      "get": {
        "description": "Every channel and every lead move, merged, newest first.",
        "operationId": "customer_timeline_v1_customers__customer_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelinePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customer Timeline",
        "tags": [
          "customers"
        ]
      }
    },
...
  }
}